Move token generating to password.js
